:fire: resolve conflictsDev liu (#6)
* feat: remove - package.json / husky

删除了 husky 的多余配置

* feat: :new: init map

地图初始化，标点基础demo实现，点击弹出框基础实现

* feat: :new: add animation to map punctuation movement and click zoomIn 1

将原有的点击标点用echarts功能直接移动，修改为用百度地图api属性动画式移动，同时当移动完成后，再进行地图下钻

* refactor: :fire: resolve conflicts

解决文件冲突问题
