+++ public/index.html
@@ -14,9 +14,27 @@
         padding: 0;
         margin: 0;
       }
+
+      .BMap_cpyCtrl {
+        display: none;
+      }
+
+      .anchorBL a {
+        display: none;
+      }
+
+      .BMap_scaleCtrl {
+        bottom: 6px !important;
+        left: 6px !important;
+      }
     </style>
     <title><%= htmlWebpackPlugin.options.title %></title>
+    <script
+      type="text/javascript"
+      src="https://api.map.baidu.com/api?v=3.0&ak=IoutUGo6SaFtmhnWZ8cSaPda74aRyPlT"
+    ></script>
   </head>
+
   <body>
     <noscript>
       <strong
