+++ webpack.base.conf.js
@@ -1,0 +1,5 @@
+module.exports = {
+  externals: {
+    BMap: "BMap",
+  },
+};
